separate blocks for teaser and detailed teaser
