+++ blocks/carousel/carousel.css
@@ -27,12 +27,14 @@
 }
 
 /* teaser inside carousel have same height and texts are centered */
+.carousel.block > .panel-container > .detailed-teaser.block,
 .carousel.block > .panel-container > .teaser.block {
   flex-shrink: 0;
   scroll-snap-align: start;
   height:auto;
 }
 
+.carousel.block > .panel-container > .detailed-teaser.block > .foreground,
 .carousel.block > .panel-container > .teaser.block > .foreground {
   height: 100%;
   display: flex;
