+++ blocks/carousel/carousel.js
@@ -1,3 +1,4 @@
+import { generateDetailedTeaserDOM } from '../detailed-teaser/detailed-teaser.js';
 import { generateTeaserDOM } from '../teaser/teaser.js';
 
 // callback for touch based scrolling event
@@ -32,15 +33,19 @@ export default function decorate(block) {
   // get all children elements
   const panels = [...block.children];
 
-  // loop through all teaser blocks
+  // loop through all children blocks
   [...panels].forEach((panel, i) => {
-    // generate the teaser panel
+    // generate the  panel
     const [image, classList, ...rest] = panel.children;
     const classesText = classList.textContent.trim();
     const classes = (classesText ? classesText.split(',') : []).map((c) => c && c.trim()).filter((c) => !!c);
-    const teaserDOM = generateTeaserDOM([image, ...rest], classes);
+    const blockType = [...classes].includes('detailed-teaser') ? 'detailed-teaser':'teaser';
+    // check if we have to render teaser or a detailed teaser
+    const teaserDOM = (blockType === 'detailed-teaser')
+      ? generateDetailedTeaserDOM([image,...rest], classes)
+      : generateTeaserDOM([image,...rest], classes);
     panel.textContent = '';
-    panel.classList.add('teaser', 'block');
+    panel.classList.add(blockType,'block');
     classes.forEach((c) => panel.classList.add(c.trim()));
     panel.dataset.panel = `panel_${i}`;
     panel.append(teaserDOM);
