+++ blocks/teaser/teaser.js
@@ -1,4 +1,4 @@
-function decorateButtons(...buttons) {
+export function decorateButtons(...buttons) {
   return buttons
     .map((div) => {
       const a = div.querySelector('a');
@@ -13,12 +13,11 @@ function decorateButtons(...buttons) {
     .join('');
 }
 
-function generateDefaultTeaser(props, classes) {
+export function generateTeaserDOM(props, classes) {
   // Extract properties, always same order as in model, empty string if not set
   const [pictureContainer, eyebrow, title, longDescr, shortDescr, firstCta, secondCta] = props;
   const picture = pictureContainer.querySelector('picture');
   const hasShortDescr = shortDescr.textContent.trim() !== '';
-
   // Build DOM
   const teaserDOM = document.createRange().createContextualFragment(`
     <div class='background'>${picture ? picture.outerHTML : ''}</div>
@@ -49,39 +48,6 @@ function generateDefaultTeaser(props, classes) {
 
   // add final teaser DOM and classes if used as child component
   return teaserDOM;
-}
-
-// eslint-disable-next-line no-unused-vars
-function generateDetailedTeaser(props, classes) {
-  // Extract properties, always same order as in model, empty string if not set
-  const [backImage, eyebrowContent, title, description, subjectImage, firstCta, secondCta] = props;
-  const backPicture = backImage.querySelector('picture');
-  const subjectPicture = subjectImage.querySelector('picture');
-  // Build DOM
-  const teaserDOM = document.createRange().createContextualFragment(`
-    <div class='background'>${backPicture ? backPicture.outerHTML : ''}</div>
-    <div class='foreground'>
-      <div class='text'>
-        ${eyebrowContent.innerHTML}
-        <div class='title'>${title.innerHTML}</div>
-        <div class='long-description'>${description.innerHTML}</div>
-        <div class='cta'>${decorateButtons(firstCta, secondCta)}</div>
-      </div>
-      <div class='spacer'>
-        ${subjectPicture ? subjectPicture.outerHTML : ''}
-      </div>
-    </div>
-  `);
-  // add final teaser DOM and classes if used as child component
-  return teaserDOM;
-}
-
-/* eslint-disable no-plusplus */
-export function generateTeaserDOM(props, classes) {
-  // check if we have to render teaser or a detailed teaser
-  return [...classes].includes('detailed-teaser')
-    ? generateDetailedTeaser(props, classes)
-    : generateDefaultTeaser(props, classes);
 }
 
 export default function decorate(block) {
