+++ component-definition.json
@@ -208,9 +208,9 @@
               "page": {
                 "resourceType": "core/franklin/components/block/v1/block",
                 "template": {
-                  "name": "Teaser",
+                  "name": "Detailed Teaser",
                   "model": "detailed-teaser",
-                  "fileReference": "/content/dam/default-teaser-banner.png",
+                  "fileReference": "/content/dam/adobe-summit-background.jpeg",
                   "eyebrow_logo": "/content/dam/summit-logo.png",
                   "eyebrow_title": "Eyebrow title",
                   "eyebrow_subtitle": "Eyebrow subtitle",
@@ -218,7 +218,7 @@
                   "titleType": "h3",
                   "description": "Description",
                   "subject": "/content/dam/summit-subject.png",
-                  "classes": ["detailed-teaser", "light", "left"]
+                  "classes": ["light", "left"]
                 }
               }
             }
